requirements.txt
# For sudoku solver
colorlog==6.8.2

# For Dnet
matplotlib==3.8.2
opencv-python==4.9.0.80
pandas==2.1.4
pyyaml==6.0.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
albumentations==1.4.6

